+++ assets/js/checkouts/super-token/errors/super-token-error-constants.js
@@ -51,3 +51,7 @@ const MPSuperTokenSkippableErrorMessages = [
     'authenticator flow is not supported',
 ];
 
+
+window.MPSuperTokenErrorCodes = MPSuperTokenErrorCodes;
+window.MPSuperTokenExpectedErrors = MPSuperTokenExpectedErrors;
+window.MPSuperTokenSkippableErrorMessages = MPSuperTokenSkippableErrorMessages;