+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v8.7.14 (26/02/2026)
+### Changed
+- Using fast payment flow with super token bundle enabled
+- Improve fast payment flow loading and rendering handling
+
 ## v8.7.13 (20/02/2026)
 ### Fixed
 - Remove unnecessary call to createPayment method on fast payment flow